# internal/api/openapi.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for CreateUploadBatchRequestMode.
const (
	File   CreateUploadBatchRequestMode = "file"
	Folder CreateUploadBatchRequestMode = "folder"
)

// CreateUploadBatchRequest defines model for CreateUploadBatchRequest.
type CreateUploadBatchRequest struct {
	// Mode Import mode
	Mode CreateUploadBatchRequestMode `json:"mode"`

	// Source Client source (e.g. tauri)
	Source string `json:"source"`
}

// CreateUploadBatchRequestMode Import mode
type CreateUploadBatchRequestMode string

// CreateUploadBatchResponse defines model for CreateUploadBatchResponse.
type CreateUploadBatchResponse struct {
	// BatchId Upload batch identifier
	BatchId string `json:"batch_id"`
}

// Error defines model for Error.
type Error struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// ListRawDocumentsResponse defines model for ListRawDocumentsResponse.
type ListRawDocumentsResponse struct {
	Documents []RawDocument `json:"documents"`
	Total     int           `json:"total"`
}

// RawDocument defines model for RawDocument.
type RawDocument struct {
	// CandidateModule Candidate module inferred by system
	CandidateModule string    `json:"candidate_module"`
	Confidence      float32   `json:"confidence"`
	CreatedAt       time.Time `json:"created_at"`

	// DocCategory Document category inferred by system
	DocCategory  string `json:"doc_category"`
	DocId        string `json:"doc_id"`
	OriginalPath string `json:"original_path"`
	Title        string `json:"title"`
}

// UploadFileRequest defines model for UploadFileRequest.
type UploadFileRequest struct {
	// DocTitle Optional document title override
	DocTitle *string `json:"doc_title,omitempty"`

	// File File content
	File openapi_types.File `json:"file"`

	// RelativePath Relative path inside the imported folder.
	// Must be provided when importing folders.
	RelativePath *string `json:"relative_path,omitempty"`
}

// UploadFileResponse defines model for UploadFileResponse.
type UploadFileResponse struct {
	// DocId Generated document identifier
	DocId string `json:"doc_id"`

	// StoredUri Storage URI of the uploaded file
	StoredUri string `json:"stored_uri"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse = Error

// ListRawDocumentsParams defines parameters for ListRawDocuments.
type ListRawDocumentsParams struct {
	BatchId *string `form:"batch_id,omitempty" json:"batch_id,omitempty"`
	Limit   *int    `form:"limit,omitempty" json:"limit,omitempty"`
	Offset  *int    `form:"offset,omitempty" json:"offset,omitempty"`
}

// CreateUploadBatchJSONRequestBody defines body for CreateUploadBatch for application/json ContentType.
type CreateUploadBatchJSONRequestBody = CreateUploadBatchRequest

// UploadFileToBatchMultipartRequestBody defines body for UploadFileToBatch for multipart/form-data ContentType.
type UploadFileToBatchMultipartRequestBody = UploadFileRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List raw documents
	// (GET /api/raw-documents)
	ListRawDocuments(c *gin.Context, params ListRawDocumentsParams)
	// Create a new upload batch
	// (POST /api/uploads)
	CreateUploadBatch(c *gin.Context)
	// Upload a file to an upload batch
	// (POST /api/uploads/{batch_id}/files)
	UploadFileToBatch(c *gin.Context, batchId string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// ListRawDocuments operation middleware
func (siw *ServerInterfaceWrapper) ListRawDocuments(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListRawDocumentsParams

	// ------------- Optional query parameter "batch_id" -------------

	err = runtime.BindQueryParameter("form", true, false, "batch_id", c.Request.URL.Query(), &params.BatchId)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter batch_id: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", c.Request.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", c.Request.URL.Query(), &params.Offset)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter offset: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListRawDocuments(c, params)
}

// CreateUploadBatch operation middleware
func (siw *ServerInterfaceWrapper) CreateUploadBatch(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateUploadBatch(c)
}

// UploadFileToBatch operation middleware
func (siw *ServerInterfaceWrapper) UploadFileToBatch(c *gin.Context) {

	var err error

	// ------------- Path parameter "batch_id" -------------
	var batchId string

	err = runtime.BindStyledParameterWithOptions("simple", "batch_id", c.Param("batch_id"), &batchId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter batch_id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UploadFileToBatch(c, batchId)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/api/raw-documents", wrapper.ListRawDocuments)
	router.POST(options.BaseURL+"/api/uploads", wrapper.CreateUploadBatch)
	router.POST(options.BaseURL+"/api/uploads/:batch_id/files", wrapper.UploadFileToBatch)
}